requirements.txt
streamlit
hopsworks
joblib
pandas
scikit-learn
python-dotenv